requirements.txt
python-telegram-bot[fast]~=20.3
fastapi~=0.115.1
uvicorn~=0.22.0
databases~=0.9.0
asyncpg~=0.29.0
SQLAlchemy~=2.0.39
pydantic[dotenv]==1.10.9
